Changed count_type to 16-bit to avoid integer overflow

+++ AMSimulationDataFormats/interface/Helper.h
@@ -12,8 +12,8 @@ namespace slhcl1tt {
 
 // Typically, short is 16 bits, int is 32 bits, long is 32 bits and long long is 64 bits
 // But we need to guarantee on the size of the integer
-typedef uint8_t  count_type; // for frequency: 0 - 255
-typedef uint16_t bit_type;   // for DC bits: 0 - 2^16 (=65536 max for nDCBits_=4)
+typedef uint16_t count_type; // for frequency: 0 - 2^16-1 (=65535 max)
+typedef uint16_t bit_type;   // for DC bits: 0 - 2^16-1 (=65535 max for nDCBits_=4)
 typedef uint32_t id_type;    // generic
 typedef uint32_t addr_type;  // for full superstrip id, now reduced to 32-bit (from 64-bit) and is therefore the same as id_type
 typedef uint32_t key_type;   // for key used in the pattern matching map, now increased to 32-bit (from 16-bit) and is therefore the same as id_type
